+++ effectful/CHANGELOG.md
@@ -3,6 +3,7 @@
   `Eff` for compatibility with existing code.
 * Re-export `writeTMVar` from `stm-2.5.1.0` in `Effectful.Concurrent.STM`.
 * Add `cancelMany` to `Effectful.Concurrent.Async`.
+* Add `Effectful.ST` and `Effectful.ST.STRef`
 
 # effectful-core-2.6.0.0 (2025-06-13)
 * Adjust `generalBracket` with `base >= 4.21` to make use of the new exception